Refactor discord for list gate command
We needed to improve the structure of the discord module to make it more
easy to extend it's functionality and make it safer with validation.
The new `list_gates` commands works but visibility might be improved
later.

+++ backend/src/cli.rs
@@ -220,7 +220,7 @@ pub enum GateCmd {
         domain_id: u64,
         /// The percentage of reputation needed to get the role
         #[clap(value_hint = ValueHint::Other)]
-        reputation: u32,
+        reputation: u8,
         /// The discord role id
         #[clap(value_hint = ValueHint::Other)]
         role_id: u64,
@@ -238,7 +238,7 @@ pub enum GateCmd {
         domain_id: u64,
         /// The percentage of reputation needed to get the role
         #[clap(value_hint = ValueHint::Other)]
-        reputation: u32,
+        reputation: u8,
         /// The discord role id
         #[clap(value_hint = ValueHint::Other)]
         role_id: u64,
